requirement.txt
streamlit
pandas
request
plotly
numpy